[Block Conversion]: Show `group` option in `Group` blocks
